refactore: 페이지 구조 개선 및 새로운 컴포넌트 추가
