+++ src/pages/Analytics/components/index.js
@@ -1,0 +1,3 @@
+export { CommentSentiment } from './CommentSentiment';
+export { DateRangeSelector } from './DateRangeSelector';
+export { createStatCard } from './StatCardCreator';
