+++ src/pages/ContentManagement/components/index.js
@@ -1,0 +1,2 @@
+export { SearchFilter } from './SearchFilter';
+export { TabMenu } from './TabMenu';
