+++ src/pages/Dashboard/components/index.js
@@ -1,0 +1,3 @@
+export { ActivityItem } from './ActivityItem';
+export { createStatCard } from './StatCardCreator';
+export { TrendSection } from './TrendSection';
