+++ src/pages/Login/components/KakaoIcon.jsx
@@ -1,0 +1,11 @@
+export const KakaoIcon = ({ size = 16, className = "" }) => (
+  <svg 
+    width={size} 
+    height={size} 
+    viewBox="0 0 24 24" 
+    className={className}
+    fill="currentColor"
+  >
+    <path d="M12 3c5.799 0 10.5 3.664 10.5 8.185 0 4.52-4.701 8.184-10.5 8.184a13.5 13.5 0 0 1-1.727-.11l-4.408 2.883c-.501.265-.678.236-.472-.413l.892-3.678c-2.88-1.46-5.183-3.668-5.183-6.866C1.5 6.665 6.201 3 12 3z"/>
+  </svg>
+);
