+++ src/pages/Login/components/index.js
@@ -1,0 +1,2 @@
+export { GoogleIcon } from './GoogleIcon';
+export { KakaoIcon } from './KakaoIcon';
