+++ src/pages/Register/components/AgeOptions.jsx
@@ -1,0 +1,7 @@
+export const ageOptions = [
+  { value: "10s", label: "10대" },
+  { value: "20s", label: "20대" },
+  { value: "30s", label: "30대" },
+  { value: "40s", label: "40대" },
+  { value: "50s", label: "50대 이상" }
+];
