+++ src/pages/Register/components/index.js
@@ -1,0 +1,1 @@
+export { ageOptions } from './AgeOptions';
