+++ src/pages/SnsIntegration/components/index.js
@@ -1,0 +1,2 @@
+export { ConnectedAccounts } from './ConnectedAccounts';
+export { ScheduledPosts } from './ScheduledPosts';
