+++ src/pages/StoreManagement/components/index.js
@@ -1,0 +1,2 @@
+export { AddStoreForm } from './AddStoreForm';
+export { StoreTable } from './StoreTable';
