+++ src/pages/index.js
@@ -1,0 +1,8 @@
+export { Analytics } from './Analytics/Analytics';
+export { ContentCreation } from './ContentCreation/ContentCreation';
+export { ContentManagement } from './ContentManagement/ContentManagement';
+export { Dashboard } from './Dashboard/Dashboard';
+export { Login } from './Login/Login';
+export { Register } from './Register/Register';
+export { SnsIntegration } from './SnsIntegration/SnsIntegration';
+export { StoreManagement } from './StoreManagement/StoreManagement';
